requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
moviepy==1.0.3
pillow==10.3.0
gtts==2.5.1
python-multipart==0.0.9
jinja2==3.1.3
